refactor(web): one flows shell for setup and join
Drop PreAppShell. /setup lives in the flows group. Logo
stays unlinked only on /setup.
