+++ apps/web/src/app/(flows)/setup/page.tsx
